+++ CHANGELOG.md
@@ -1,24 +1,34 @@
-## 0.6.0 (2024-10-08)
+## [0.6.0](https://github.com/jersou/clite-parser/compare/0.5.0...0.6.0) (2024-10-09)
 
-- refactor: use @std/cli
+- feat: use [@std/cli](https://jsr.io/@std/cli/doc/parse-args) ! based on
+  [minimist](https://github.com/minimistjs/minimist).
 - feat: add alias decorator & _<prop>_alias
+- feat: add types decorator & _<prop>_types
+- feat: add defaultHelp decorator & _<prop>_default
+- feat: update help format
 
-## 0.5.0 (2024-09-24)
+## [0.5.0](https://github.com/jersou/clite-parser/compare/0.4.0...0.5.0) (2024-09-24)
 
 - doc : update, clean
 - refactor: use _help instead of _desc
 - feat: add @help decorator
+  ([c707ce7](https://github.com/jersou/clite-parser/commit/c707ce75efee6649d36a925787b4536c2ccf9d8a))
 - feat: supports plain objects
+  ([f0293da](https://github.com/jersou/clite-parser/commit/f0293dac2f0f889d0279885c64bfdc09ea242a67))
 
 ## 0.4.0 (2024-09-23)
 
 - feat: add "meta" config
+  ([ca5f347](https://github.com/jersou/clite-parser/commit/ca5f347c10f444f4a2b961c0633183af36e3ed65))
 - feat: add mainFile to config
+  ([6ebebf6](https://github.com/jersou/clite-parser/commit/6ebebf6ebd4799aef4217cd83140a500bf470a54))
 - feat: add printHelpOnError
+  ([18699a0](https://github.com/jersou/clite-parser/commit/18699a043b993b07f52dd2e11f86b91dd23fbec8))
 - doc: add MIT license
 - doc: refactor & add Contributors
 - refactor: update dep, remove imports from deno.json
 - fix: fix dcpps example
+  ([d8dd2c2](https://github.com/jersou/clite-parser/commit/d8dd2c2ecb339b211e898831f77aeba2be5c09bd))
 
 ## 0.3.2 (2024-09-08)